fix socket
