+++ src/components/index.ts
@@ -27,7 +27,6 @@ export * from './TextArea';
 export * from './UserAccountLayout';
 export * from './MenuNotification';
 export * from './MenuActions';
-export * from './NotificationContainer';
 export * from './TournamentLayout';
 export * from './UploadImage';
 export * from './SetScheduleBox';
