+++ src/hooks/index.ts
@@ -5,5 +5,4 @@ export * from './usePaymentForm';
 export * from './useVenueMutation';
 export * from './useVenueByUser';
 export * from './useDebounce';
-export * from './useSocket';
 export * from './useLocalStorage';
