--- src/hooks/useSocket.ts
@@ -1,8 +1,0 @@
-import { useContext } from 'react';
-import { SocketContext } from '@/App';
-
-export const useSocket = () => {
-  const socket = useContext(SocketContext);
-
-  return { socket };
-};
